populate owner

+++ controllers/cards.js
@@ -13,20 +13,21 @@ module.exports.getCards = (req, res, next) => {
 module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
 
-  Card.create({ name, link, owner: req.user })
+  Card.create({ name, link, owner: req.user }).populate('owner')
     .then((card) => res.send(card))
     .catch(next);
 };
 
 module.exports.deleteCardById = (req, res, next) => {
-  Card.findById(req.params.cardId)
+  Card.findById(req.params.cardId).populate('owner')
     .then((card) => {
       if (!card) return Promise.reject(notFoundError);
       if (card.owner.toString() !== req.user._id) return Promise.reject(forbiddenError);
 
       return card;
     })
-    .then(() => Card.findByIdAndRemove(req.params.cardId, { runValidators: true }))
+    .then(() => Card.findByIdAndRemove(req.params.cardId, { runValidators: true }).populate('owner'))
+
     .then((card) => {
       if (!card) return Promise.reject(notFoundError);
       return res.send(card);
@@ -40,6 +41,7 @@ module.exports.likeCard = (req, res, next) => {
     { $addToSet: { likes: req.user._id } },
     { new: true, runValidators: true },
   )
+    .populate('owner')
     .then((card) => {
       if (!card) return Promise.reject(notFoundError);
       return res.send(card);
@@ -53,6 +55,7 @@ module.exports.dislikeCard = (req, res, next) => {
     { $pull: { likes: req.user._id } },
     { new: true, runValidators: true },
   )
+    .populate('owner')
     .then((card) => {
       if (!card) return Promise.reject(notFoundError);
       return res.send(card);
